contract: Merge all 3 tests into 1 file
